# openapi.yaml
openapi: 3.0.0
info:
  x-skaffolder-id-project: 5ea9987e82f82d390e780496
  title: techhubvirtual
  description: You can obtain JWT token from login API
  version: 0.1
servers:
  - url: http://localhost:3000/api
    description: Local server for testing
paths:
...
  "/freelancer/{id}":
    delete:
      x-skaffolder-id: 5ebed53aa940665206b00bb2
      x-skaffolder-name: delete
      x-skaffolder-id-resource: 5ebeafafa940665206b0088b
      x-skaffolder-resource: Freelancer
      x-skaffolder-crudAction: delete
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION delete
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION delete
      tags:
        - Freelancer
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      x-skaffolder-id: 5ebed53aa940665206b00bad
      x-skaffolder-name: get
      x-skaffolder-id-resource: 5ebeafafa940665206b0088b
      x-skaffolder-resource: Freelancer
      x-skaffolder-crudAction: get
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION get
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION get
      tags:
        - Freelancer
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Freelancer"
    post:
      x-skaffolder-id: 5ebed53aa940665206b00bab
      x-skaffolder-name: update
      x-skaffolder-id-resource: 5ebeafafa940665206b0088b
      x-skaffolder-resource: Freelancer
      x-skaffolder-crudAction: update
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION update
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION update
      tags:
        - Freelancer
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Freelancer"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Freelancer"
...
  "/investor/{id}":
    delete:
      x-skaffolder-id: 5ebed5a1a940665206b00bca
      x-skaffolder-name: delete
      x-skaffolder-id-resource: 5ebeafb7a940665206b0088f
      x-skaffolder-resource: Investor
      x-skaffolder-crudAction: delete
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION delete
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION delete
      tags:
        - Investor
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      x-skaffolder-id: 5ebed5a1a940665206b00bc7
      x-skaffolder-name: get
      x-skaffolder-id-resource: 5ebeafb7a940665206b0088f
      x-skaffolder-resource: Investor
      x-skaffolder-crudAction: get
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION get
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION get
      tags:
        - Investor
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Investor"
    post:
      x-skaffolder-id: 5ebed5a1a940665206b00bc4
      x-skaffolder-name: update
      x-skaffolder-id-resource: 5ebeafb7a940665206b0088f
      x-skaffolder-resource: Investor
      x-skaffolder-crudAction: update
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION update
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION update
      tags:
        - Investor
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Investor"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Investor"
...
  "/mentor/{id}":
    delete:
      x-skaffolder-id: 5ebed542a940665206b00bbc
      x-skaffolder-name: delete
      x-skaffolder-id-resource: 5ebeafa0a940665206b00887
      x-skaffolder-resource: Mentor
      x-skaffolder-crudAction: delete
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION delete
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION delete
      tags:
        - Mentor
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      x-skaffolder-id: 5ebed542a940665206b00bb5
      x-skaffolder-name: get
      x-skaffolder-id-resource: 5ebeafa0a940665206b00887
      x-skaffolder-resource: Mentor
      x-skaffolder-crudAction: get
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION get
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION get
      tags:
        - Mentor
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Mentor"
    post:
      x-skaffolder-id: 5ebed542a940665206b00bb7
      x-skaffolder-name: update
      x-skaffolder-id-resource: 5ebeafa0a940665206b00887
      x-skaffolder-resource: Mentor
      x-skaffolder-crudAction: update
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION update
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION update
      tags:
        - Mentor
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Mentor"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Mentor"
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Freelancer:
      x-skaffolder-id: 5ebeafafa940665206b0088b
      x-skaffolder-id-db: 5ea9987e82f82d390e780497
      x-skaffolder-id-entity: 5ebeafaea940665206b0088a
      x-skaffolder-url: /freelancer
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        Email:
          type: number
          x-skaffolder-id-attr: 5ebed52fa940665206b00ba8
          x-skaffolder-type: Number
        Name:
          type: string
          x-skaffolder-id-attr: 5ebed52fa940665206b00ba7
          x-skaffolder-type: String
          x-skaffolder-required: true
        Speciality:
          type: string
          x-skaffolder-id-attr: 5ebed52fa940665206b00ba9
          x-skaffolder-type: String
      x-skaffolder-relations: 
      required:
        - _id
        - Name
    Investor:
      x-skaffolder-id: 5ebeafb7a940665206b0088f
      x-skaffolder-id-db: 5ea9987e82f82d390e780497
      x-skaffolder-id-entity: 5ebeafb7a940665206b0088e
      x-skaffolder-url: /investor
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        Cell:
          type: number
          x-skaffolder-id-attr: 5ebed572a940665206b00bbf
          x-skaffolder-type: Number
          x-skaffolder-required: true
        Email:
          type: string
          x-skaffolder-id-attr: 5ebed572a940665206b00bc0
          x-skaffolder-type: String
          x-skaffolder-required: true
        Name:
          type: string
          x-skaffolder-id-attr: 5ebed572a940665206b00bbe
          x-skaffolder-type: String
          x-skaffolder-required: true
      x-skaffolder-relations: 
      required:
        - _id
        - Cell
        - Email
        - Name
    Mentor:
      x-skaffolder-id: 5ebeafa0a940665206b00887
      x-skaffolder-id-db: 5ea9987e82f82d390e780497
      x-skaffolder-id-entity: 5ebeafa0a940665206b00886
      x-skaffolder-url: /mentor
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        cellnumber:
          type: number
          x-skaffolder-id-attr: 5ebeb462a940665206b008c0
          x-skaffolder-type: Number
          x-skaffolder-required: true
        description:
          type: string
          x-skaffolder-id-attr: 5ebeb462a940665206b008c1
          x-skaffolder-type: String
          x-skaffolder-required: true
        name:
          type: string
          x-skaffolder-id-attr: 5ebeb462a940665206b008bf
          x-skaffolder-type: String
          x-skaffolder-required: true
      x-skaffolder-relations: 
      required:
        - _id
        - cellnumber
        - description
        - name
...